Refine app detail architecture
