+++ app/src/main/kotlin/sk/styk/martin/apkanalyzer/dependencyinjection/ApplicationModule.kt
@@ -1,8 +1,5 @@
 package sk.styk.martin.apkanalyzer.dependencyinjection
 
-import android.app.Application
-import android.app.NotificationManager
-import android.content.ContentResolver
 import android.content.Context
 import com.google.firebase.analytics.FirebaseAnalytics
 import dagger.Module
@@ -20,13 +17,6 @@ class ApplicationModule {
     @Provides
     @Singleton
     fun provideApplicationScope(): CoroutineScope = CoroutineScope(SupervisorJob())
-
-    @Provides
-    @Singleton
-    fun provideContentResolver(application: Application): ContentResolver = application.contentResolver
-
-    @Provides
-    fun provideNotificationManager(@ApplicationContext context: Context): NotificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
 
     @Provides
     fun provideFirebaseAnalytics(@ApplicationContext context: Context): FirebaseAnalytics = FirebaseAnalytics.getInstance(context)
