--- app/src/main/kotlin/sk/styk/martin/apkanalyzer/util/OutputFilePickerRequest.kt
@@ -1,3 +1,0 @@
-package sk.styk.martin.apkanalyzer.util
-
-data class OutputFilePickerRequest(val fileName: String, val fileType: String)
