--- app/src/main/kotlin/sk/styk/martin/apkanalyzer/util/file/DrawableExtensions.kt
@@ -1,29 +1,0 @@
-package sk.styk.martin.apkanalyzer.util.file
-
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.Drawable
-
-fun Drawable.toBitmap(): Bitmap {
-    val bitmap: Bitmap?
-
-    if (this is BitmapDrawable) {
-        val bitmapDrawable = this
-        if (bitmapDrawable.bitmap != null) {
-            return bitmapDrawable.bitmap
-        }
-    }
-
-    bitmap =
-        if (this.intrinsicWidth <= 0 || this.intrinsicHeight <= 0) {
-            Bitmap.createBitmap(1, 1, Bitmap.Config.ARGB_8888) // Single color bitmap will be created of 1x1 pixel
-        } else {
-            Bitmap.createBitmap(this.intrinsicWidth, this.intrinsicHeight, Bitmap.Config.ARGB_8888)
-        }
-
-    val canvas = Canvas(bitmap)
-    this.setBounds(0, 0, canvas.width, canvas.height)
-    this.draw(canvas)
-    return bitmap
-}
