+++ core/apps/src/main/kotlin/sk/styk/martin/apkanalyzer/core/apps/model/BroadcastReceiver.kt
@@ -1,3 +1,6 @@
 package sk.styk.martin.apkanalyzer.core.apps.model
 
 data class BroadcastReceiver(val name: String, val permission: String? = null, val isExported: Boolean = false)
+
+val BroadcastReceiver.isExternallyReachableWithoutPermission: Boolean
+    get() = isExported && permission.isNullOrBlank()
