+++ core/common/src/main/kotlin/sk/styk/martin/apkanalyzer/core/common/coroutines/RunCatching.kt
@@ -2,10 +2,8 @@ package sk.styk.martin.apkanalyzer.core.common.coroutines
 
 import kotlin.coroutines.cancellation.CancellationException
 
-suspend inline fun <T> runCatchingCancellable(block: () -> T): Result<T> = try {
-    Result.success(block())
-} catch (exception: CancellationException) {
-    throw exception
-} catch (throwable: Throwable) {
-    Result.failure(throwable)
+suspend inline fun <T> runCatchingCancellable(block: () -> T): Result<T> = runCatching(block).onFailure {
+    if (it is CancellationException) {
+        throw it
+    }
 }
