+++ feature/app-detail/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/appdetail/impl/manifest/ManifestViewModel.kt
@@ -6,13 +6,18 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import sk.styk.martin.apkanalyzer.core.apps.analysis.ManifestParser
+import sk.styk.martin.apkanalyzer.core.apps.analysis.ParsedManifest
 import sk.styk.martin.apkanalyzer.core.common.coroutines.DispatcherProvider
 import sk.styk.martin.apkanalyzer.core.common.logger.Logger
 import sk.styk.martin.apkanalyzer.feature.appdetail.api.AppDetailInput
@@ -28,12 +33,18 @@ internal class ManifestViewModel @AssistedInject constructor(@Assisted private v
         fun create(target: AppDetailInput): ManifestViewModel
     }
 
-    private val field = MutableStateFlow<ManifestState>(ManifestState.Loading)
-    val state: StateFlow<ManifestState> = field
+    private val source = MutableStateFlow<ManifestSource>(ManifestSource.Loading)
+    private val query = MutableStateFlow("")
 
-    private var manifest = ""
-    private var additionalInstalledSplits = 0
-    private var searchJob: Job? = null
+    val state: StateFlow<ManifestState> = combine(source, query) { source, query ->
+        when (source) {
+            ManifestSource.Loading -> ManifestState.Loading
+            ManifestSource.Error -> ManifestState.Error
+            is ManifestSource.Ready -> source.filteredBy(query)
+        }
+    }
+        .flowOn(dispatcherProvider.default())
+        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), ManifestState.Loading)
 
     init {
         loadManifest()
@@ -42,97 +53,90 @@ internal class ManifestViewModel @AssistedInject constructor(@Assisted private v
     fun onAction(action: ManifestAction) {
         when (action) {
             ManifestAction.Retry -> loadManifest()
-            is ManifestAction.ChangeQuery -> updateQuery(action.query)
+            is ManifestAction.ChangeQuery -> query.value = action.query
         }
     }
 
     private fun loadManifest() {
-        field.value = ManifestState.Loading
+        source.value = ManifestSource.Loading
+        query.value = ""
         viewModelScope.launch {
             val result = manifestParser.manifest(appDetailInput.toAppReference())
-            field.value = result.onFailure {
-                Logger.e(TAG, it, "Can not load manifest for $appDetailInput")
-            }.fold(
-                onSuccess = {
-                    manifest = it.xml
-                    additionalInstalledSplits = it.additionalInstalledSplits
-                    it.xml.toState("")
-                },
-                onFailure = { ManifestState.Error },
-            )
-        }
-    }
-
-    private fun updateQuery(query: String) {
-        if (manifest.isEmpty()) return
-        (field.value as? ManifestState.Loaded)?.let { current ->
-            field.value = current.copy(query = query)
-        }
-        searchJob?.cancel()
-        searchJob = viewModelScope.launch {
-            val filteredState = withContext(dispatcherProvider.default()) {
-                manifest.toState(query)
-            }
-            if ((field.value as? ManifestState.Loaded)?.query == query) {
-                field.value = filteredState
-            }
-        }
-    }
-
-    private fun String.toState(query: String): ManifestState.Loaded {
-        val lines = lineSequence().toList()
-        val matchingIndices = if (query.isBlank()) {
-            lines.indices.toList()
-        } else {
-            lines.indices.filter { index -> lines[index].contains(query, ignoreCase = true) }
-        }
-        val displayedIndices = if (query.isBlank()) {
-            matchingIndices
-        } else {
-            buildSet {
-                matchingIndices.forEach { index ->
-                    addAll(lines.ownerElementRange(index))
-                    add(index)
-                }
-            }.sorted()
-        }
-        val matchingIndexSet = matchingIndices.toSet()
-        return ManifestState.Loaded(
-            query = query,
-            displayedLines = displayedIndices.map { index ->
-                ManifestLine(
-                    number = index + 1,
-                    text = lines[index],
-                    isMatch = query.isNotBlank() && index in matchingIndexSet,
+            source.value = withContext(dispatcherProvider.default()) {
+                result.onFailure {
+                    Logger.e(TAG, it, "Can not load manifest for $appDetailInput")
+                }.fold(
+                    onSuccess = { it.toSource() },
+                    onFailure = { ManifestSource.Error },
                 )
-            }.toImmutableList(),
-            lineCount = lines.size,
-            matchCount = if (query.isBlank()) 0 else matchingIndices.size,
-            additionalInstalledSplits = additionalInstalledSplits,
-        )
-    }
-
-    private fun List<String>.ownerElementRange(index: Int): IntRange {
-        val matchIndent = get(index).leadingSpaceCount()
-        val matchingLine = get(index).trimStart()
-        val ownerIndex = if (matchingLine.startsWith("<") && !matchingLine.startsWith("</")) {
-            index
-        } else {
-            (index downTo 0).firstOrNull { candidateIndex ->
-                val candidate = get(candidateIndex)
-                val trimmed = candidate.trimStart()
-                candidate.leadingSpaceCount() < matchIndent &&
-                    trimmed.startsWith("<") &&
-                    !trimmed.startsWith("</")
             }
-        } ?: return index..index
-        val endIndex = (ownerIndex until size).firstOrNull { candidateIndex ->
-            get(candidateIndex).trimEnd().endsWith(">")
-        } ?: index
-        return ownerIndex..endIndex
+        }
     }
-
-    private fun String.leadingSpaceCount(): Int = indexOfFirst { !it.isWhitespace() }
-        .takeIf { it >= 0 }
-        ?: length
 }
+
+private sealed interface ManifestSource {
+    data object Loading : ManifestSource
+    data object Error : ManifestSource
+    data class Ready(val lines: ImmutableList<String>, val additionalInstalledSplits: Int) : ManifestSource
+}
+
+private fun ParsedManifest.toSource() = ManifestSource.Ready(
+    lines = xml.lineSequence().toImmutableList(),
+    additionalInstalledSplits = additionalInstalledSplits,
+)
+
+private fun ManifestSource.Ready.filteredBy(query: String): ManifestState.Loaded {
+    val matchingIndices = if (query.isBlank()) {
+        lines.indices.toList()
+    } else {
+        lines.indices.filter { index -> lines[index].contains(query, ignoreCase = true) }
+    }
+    val displayedIndices = if (query.isBlank()) {
+        matchingIndices
+    } else {
+        buildSet {
+            matchingIndices.forEach { index ->
+                addAll(lines.ownerElementRange(index))
+                add(index)
+            }
+        }.sorted()
+    }
+    val matchingIndexSet = matchingIndices.toSet()
+    return ManifestState.Loaded(
+        query = query,
+        displayedLines = displayedIndices.map { index ->
+            ManifestLine(
+                number = index + 1,
+                text = lines[index],
+                isMatch = query.isNotBlank() && index in matchingIndexSet,
+            )
+        }.toImmutableList(),
+        lineCount = lines.size,
+        matchCount = if (query.isBlank()) 0 else matchingIndices.size,
+        additionalInstalledSplits = additionalInstalledSplits,
+    )
+}
+
+private fun List<String>.ownerElementRange(index: Int): IntRange {
+    val matchIndent = get(index).leadingSpaceCount()
+    val matchingLine = get(index).trimStart()
+    val ownerIndex = if (matchingLine.startsWith("<") && !matchingLine.startsWith("</")) {
+        index
+    } else {
+        (index downTo 0).firstOrNull { candidateIndex ->
+            val candidate = get(candidateIndex)
+            val trimmed = candidate.trimStart()
+            candidate.leadingSpaceCount() < matchIndent &&
+                trimmed.startsWith("<") &&
+                !trimmed.startsWith("</")
+        }
+    } ?: return index..index
+    val endIndex = (ownerIndex until size).firstOrNull { candidateIndex ->
+        get(candidateIndex).trimEnd().endsWith(">")
+    } ?: index
+    return ownerIndex..endIndex
+}
+
+private fun String.leadingSpaceCount(): Int = indexOfFirst { !it.isWhitespace() }
+    .takeIf { it >= 0 }
+    ?: length
